add: added changes in Array
