+++ datastax/Arrays/priority_queue.py
@@ -1,7 +1,7 @@
 # Priority Queue implementation using Lists (Pseudo Arrays)
 from typing import Any, Callable
 
-from datastax.Arrays.queue import Queue
+from datastax.Arrays.Queue import Queue
 from datastax.errors import OverFlowError, UnderFlowError
